Fix issue with colleciton/view name conflict checking in cluster.
