+++ tests/js/common/shell/shell-view-arangosearch-link.js
